07_pjt_master/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.11
django-extensions==3.2.1
djangorestframework==3.13.1
idna==3.4
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
